+++ client/projects/cms/src/app/modules/login/login.module.ts
@@ -3,6 +3,7 @@ import {NgModule} from '@angular/core';
 import {ReactiveFormsModule} from '@angular/forms';
 import {MatButtonModule} from '@angular/material/button';
 import {MatCardModule} from '@angular/material/card';
+import {MatCheckboxModule} from '@angular/material/checkbox';
 import {MatDialogModule} from '@angular/material/dialog';
 import {MatFormFieldModule} from '@angular/material/form-field';
 import {MatInputModule} from '@angular/material/input';
@@ -35,6 +36,7 @@ const routes: Routes = [{
     MatInputModule,
     MatDialogModule,
     MatSelectModule,
+    MatCheckboxModule,
 
     /**
      * Ng Helpers