+++ functions/src/callable/impersonate.ts
@@ -1,12 +1,16 @@
-import {auth} from 'firebase-admin';
+import * as admin from 'firebase-admin';
 import * as functions from 'firebase-functions';
 import {STATIC_CONFIG} from '../consts/static-config.const';
 import {hasRole} from '../utils/auth';
 
-export const getUser = functions
+export const impersonate = functions
   .region(STATIC_CONFIG.cloudRegion)
   .https
   .onCall(async (data, context) => {
     hasRole(context, 'admin');
-    return auth().getUser(data);
+
+    const auth = admin.auth();
+    const user = await auth.getUser(data);
+
+    return await auth.createCustomToken(data, user.customClaims)
   });