+++ functions/src/callable/update-user.ts
@@ -1,19 +1,20 @@
 import {auth} from 'firebase-admin';
 import * as functions from 'firebase-functions';
 import {STATIC_CONFIG} from '../consts/static-config.const';
+import {hasRole} from '../utils/auth';
 
 export const updateUser = functions
   .region(STATIC_CONFIG.cloudRegion)
   .https
   .onCall(async (data, context) => {
-    if (!context.auth || !context.auth.token.role) {
-      throw new functions.https.HttpsError(
-        'failed-precondition',
-        'The function must be called ' + 'while authenticated.'
-      );
-    }
+    hasRole(context, 'admin');
 
     const {id, ...update} = data;
 
-    await auth().updateUser(id, update);
+    try {
+      await auth().updateUser(id, update);
+    } catch (e) {
+      console.error(e);
+      throw new functions.https.HttpsError('internal', e.message);
+    }
   });